+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+Nothing yet.
+
+## [1.2.0] - 2022-03-26
+
 ### Added
 
 - cfg_shares table to store data about custom shares
@@ -15,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Implemented IBootstrap
 - Updated translations from transifex
+- Updated [@nextcloud-vue](https://github.com/nextcloud/nextcloud-vue) to v5.2.1
 
 ### Fixed
 
@@ -83,7 +88,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Initial release
 
-[Unreleased]: https://github.com/jimmyl0l3c/cfg_share_links/compare/v1.1.1...HEAD
+[Unreleased]: https://github.com/jimmyl0l3c/cfg_share_links/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/jimmyl0l3c/cfg_share_links/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/jimmyl0l3c/cfg_share_links/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/jimmyl0l3c/cfg_share_links/compare/v1.0.3...v1.1.0
 [1.0.3]: https://github.com/jimmyl0l3c/cfg_share_links/compare/v1.0.2...v1.0.3